uml-test-init: create more loopback devices when loading loop module
With the upcoming addition of an always mounted bootloader slot in the
default test system config, more loopback devices are needed to support
simultanious mount of slots and bundles during test.

Ususally, it should be enough to specify the "max_loops" module
parameter, but gqthat doesn't seem to be enough when running in UML
context.

Make the devices appear in /dev/ by running `mknod` in a for loop.