test/install-fixture: add always mounted bootloader slot
In order to test the "already-mounted" feature for updating slots
mounted by the system at boot, such a "bootloader" slot is added to the
default test configuration.

Since an additional slot is added, the test for correct number of slots
in test/service.c is updated from 5 slots to 6 slots.